+++ JDI.Light/JDI.Light.Tests/Tests/Simple/DateTests.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using NUnit.Framework;
 
-namespace JDI.Light.Tests.Tests.Common
+namespace JDI.Light.Tests.Tests.Simple
 {
     [TestFixture]
     public class DateTests : TestBase
