+++ JDI.Light/JDI.Light.Tests/Tests/Simple/MonthTests.cs
@@ -4,7 +4,7 @@ using NUnit.Framework;
 namespace JDI.Light.Tests.Tests.Simple
 {
     [TestFixture]
-    public class TimeTests : TestBase
+    public class MonthTests : TestBase
     {
         private readonly DateTime _dateTime = new DateTime(2019, 4, 1, 15, 0, 0);
 
@@ -19,34 +19,35 @@ namespace JDI.Light.Tests.Tests.Simple
         [Test]
         public void GetDateTest()
         {
-            Assert.AreEqual(TestSite.Html5Page.BookingTime.Value(), "11:00");
+            Assert.AreEqual(TestSite.Html5Page.MonthDate.Value(), "2018-05");
         }
 
         [Test]
         public void MinTest()
         {
-            Assert.AreEqual(TestSite.Html5Page.BookingTime.Min(), "9:00");
+            Assert.AreEqual(TestSite.Html5Page.MonthDate.Min(), "2015-03");
         }
 
         [Test]
         public void MaxTest()
         {
-            Assert.AreEqual(TestSite.Html5Page.BookingTime.Max(), "18:00");
+            Assert.AreEqual(TestSite.Html5Page.MonthDate.Max(), "2020-12");
         }
 
         [Test]
         public void SetDateTimeTest()
         {
-            TestSite.Html5Page.BookingTime.SetDateTime("05:00");
-            Assert.AreEqual(TestSite.Html5Page.BookingTime.Value(), "05:00");
+            TestSite.Html5Page.MonthDate.SetDateTime("2018-10");
+            Assert.AreEqual(TestSite.Html5Page.MonthDate.Value(), "2018-10");
         }
 
         [Test]
         public void SetMonthTest()
         {
-            TestSite.Html5Page.BookingTime.Format = "H:mm";
-            TestSite.Html5Page.BookingTime.SetDateTime(_dateTime);
-            Assert.AreEqual(TestSite.Html5Page.BookingTime.Value(), "15:00");
+            TestSite.Html5Page.MonthDate.Format = "yyyy-MM";
+            TestSite.Html5Page.MonthDate.SetDateTime(_dateTime);
+
+            Assert.AreEqual(TestSite.Html5Page.MonthDate.Value(), "2019-04");
         }
     }
 }
