+++ JDI.Light/JDI.Light.Tests/Tests/Simple/TimeTests.cs
@@ -1,10 +1,10 @@
 ﻿using System;
 using NUnit.Framework;
 
-namespace JDI.Light.Tests.Tests.Common
+namespace JDI.Light.Tests.Tests.Simple
 {
     [TestFixture]
-    public class MonthTests : TestBase
+    public class TimeTests : TestBase
     {
         private readonly DateTime _dateTime = new DateTime(2019, 4, 1, 15, 0, 0);
 
@@ -19,35 +19,34 @@ namespace JDI.Light.Tests.Tests.Common
         [Test]
         public void GetDateTest()
         {
-            Assert.AreEqual(TestSite.Html5Page.MonthDate.Value(), "2018-05");
+            Assert.AreEqual(TestSite.Html5Page.BookingTime.Value(), "11:00");
         }
 
         [Test]
         public void MinTest()
         {
-            Assert.AreEqual(TestSite.Html5Page.MonthDate.Min(), "2015-03");
+            Assert.AreEqual(TestSite.Html5Page.BookingTime.Min(), "9:00");
         }
 
         [Test]
         public void MaxTest()
         {
-            Assert.AreEqual(TestSite.Html5Page.MonthDate.Max(), "2020-12");
+            Assert.AreEqual(TestSite.Html5Page.BookingTime.Max(), "18:00");
         }
 
         [Test]
         public void SetDateTimeTest()
         {
-            TestSite.Html5Page.MonthDate.SetDateTime("2018-10");
-            Assert.AreEqual(TestSite.Html5Page.MonthDate.Value(), "2018-10");
+            TestSite.Html5Page.BookingTime.SetDateTime("05:00");
+            Assert.AreEqual(TestSite.Html5Page.BookingTime.Value(), "05:00");
         }
 
         [Test]
         public void SetMonthTest()
         {
-            TestSite.Html5Page.MonthDate.Format = "yyyy-MM";
-            TestSite.Html5Page.MonthDate.SetDateTime(_dateTime);
-
-            Assert.AreEqual(TestSite.Html5Page.MonthDate.Value(), "2019-04");
+            TestSite.Html5Page.BookingTime.Format = "H:mm";
+            TestSite.Html5Page.BookingTime.SetDateTime(_dateTime);
+            Assert.AreEqual(TestSite.Html5Page.BookingTime.Value(), "15:00");
         }
     }
 }
