+++ JDI.Light/JDI.Light.Tests/Tests/Simple/WeekTests.cs
@@ -2,7 +2,7 @@
 using System.Globalization;
 using NUnit.Framework;
 
-namespace JDI.Light.Tests.Tests.Common
+namespace JDI.Light.Tests.Tests.Simple
 {
     [TestFixture]
     public class WeekTests : TestBase
